remove .ftp_cache: do an FTP 'dir' command instead
This should guarantee that old files are always removed properly. Also
avoids re-trying to delete files that are already gone.

FTP files expiration is checked using the timestamp in the filename.
Files of the same type (daily, weekly, monthly) as the current backup
are removed if older than the threshold value (in days).